twitterbot/requirements.txt
attrs==20.3.0
blessed==1.17.6
botocore==1.20.68
certifi==2020.12.5
chalice==1.22.4
chardet==4.0.0
click==7.1.2
idna==2.10
inquirer==2.7.0
jmespath==0.10.0
mypy-extensions==0.4.3
oauthlib==3.1.0
prompt-toolkit==3.0.18
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
PyYAML==5.4.1
readchar==2.0.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.4
wcwidth==0.2.5
wit==6.0.0
